fix data type "double" in  is not supported by "mssql" database

+++ src/core/entities/player-statistics.entity.ts
@@ -27,13 +27,13 @@ export class PlayerStatistics {
   @Column()
   passes_attempted: number;
 
-  @Column({ type: 'double' })
+  @Column({ type: 'float' })
   pass_accuracy: number;
 
   @Column()
   shots_on_target: number;
 
-  @Column({ type: 'double' })
+  @Column({ type: 'float' })
   shot_accuracy: number;
 
   @Column()
@@ -42,7 +42,7 @@ export class PlayerStatistics {
   @Column()
   dribbles_attempted: number;
 
-  @Column({ type: 'double' })
+  @Column({ type: 'float' })
   dribble_success_rate: number;
 
   @Column()
